requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
Flask-Login==0.6.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Werkzeug==2.3.4
bcrypt==4.0.1
qrcode==7.3
email-validator==1.3.1
PyMySQL==1.0.2
Pillow
stripe
flask-wtf
Flask-Mail
itsdangerous
pytest
coverage
